Added a spell upgrade preview system
Added a spell upgrade preview system for the user interface. When the player hovers over a spell during a spell upgrade event, the result of the upgrade will temporarily be displayed. Spell upgrade results that are based on chance are now predetermined as soon as the preview appears, and these results are cached until the spell changes or Bum-bo exits the game.
Encapsulated spell upgrades into their own class of objects, reducing code duplication and facilitating the development of further spell upgrades or modifications to existing upgrades.